Update AI_stats_lab.py (2024-DS-31)

+++ AI_stats_lab.py
@@ -85,9 +85,9 @@ def normal_pdf(x, mu, sigma):
         1/(sqrt(2π)σ) * exp(-(x-μ)^2 / (2σ^2))
     """
     def normal_pdf(x, mu, sigma):
-        coeff = 1 / (np.sqrt(2 * np.pi) * sigma)
-        exponent = -((x - mu) ** 2) / (2 * sigma ** 2)
-        return coeff * np.exp(exponent)
+    coeff = 1 / (np.sqrt(2 * np.pi) * float(sigma))
+    exponent = -((float(x) - float(mu)) ** 2) / (2 * float(sigma) ** 2)
+    return float(coeff * np.exp(exponent))
 
 
 def normal_histogram_analysis(mu_values,
@@ -144,14 +144,14 @@ def uniform_mean(a, b):
     """
     (a + b) / 2
     """
-    return float(a + b) / 2
+    return float(a + b) / 2.0
 
 
 def uniform_variance(a, b):
     """
     (b - a)^2 / 12
     """
-    return float((b - a) ** 2) / 12
+    return float((b - a) ** 2) / 12.0
 
 
 def uniform_histogram_analysis(a_values,
